update: fix database definitions
- Improve settings to add the missing files
- Rename more instances of "teams"
- Change database to include more of the definition files and to generate the right types of files (WIP, there's some work to do here)

The database should work with the definition files to generate data. To load the data, we are having the:

clubs.json file, which has the current teams in the game, probably a sort of copy of the clubs_def.json

players.json -> basic player definition

player_teams.json -> defines the relationship between player and their clubs/teams

squads.json -> defines the relationship between the clubs and players. List of all players from a club.

In the future we will have the championships.json and probably a match_history.json as well...

I'm looking forward to test this further, and then port all of these JSON files to a SQLite database later.